TRUE Connect Four AI (the last one was far too bad)

+++ apps/numworks_utils/src/storage/mystring.h
@@ -10,6 +10,6 @@ int strcmp(const char *s1, const char *s2);
 void *mymemcpy(void *__restrict__ dest, const void *__restrict__ src, size_t n);
 void *mymemmove(void *dest, const void *src, size_t n);
 void *mymemset(void *s, int c, size_t n);
-size_t strlen(const char *s);
+unsigned int strlen(const char *s);
 
 #endif
